refactor: extract subprocess targets to lightweight module for Windows spawn compatibility (#235)
- Create subprocess_tasks.py with calc_dir_size, scan_directory_sync,
  render_image_sync, compute_resize — no Textual/Rich imports
- Move functions from app.py, filtered_tree.py, image_preview.py
- Make MultiCursorTextArea import lazy in code_editor_helpers.py
- Fixes Windows CI: spawn context no longer imports Textual widget tree
